C#: Move the AssemblyPath class to its own file.
